+++ Source/ACE.Api.Host/Program.cs
@@ -15,13 +15,12 @@ namespace ACE.Api.Host
                 // Get the bind address and port from config:
                 var server = WebApp.Start<Startup>(url: ConfigManager.Config.ApiServer.ListenUrl);
                 Console.WriteLine($"ACE API listening at {ConfigManager.Config.ApiServer.ListenUrl}");
-                Console.ReadLine();
             }
             else
             {
                 Console.WriteLine("There was an error in your API configuration.");
-                Console.ReadLine(); // need a readline or the error flashes without being seen
             }
+            Console.ReadLine(); // need a readline or the error flashes without being seen
         }
     }
 }